working solution from existing
